Trait windows_gen::ToTokens

pub trait ToTokens {
    fn to_tokens(&self, tokens: &mut TokenStream);

    fn to_token_stream(&self) -> TokenStream { ... }

    fn into_token_stream(self) -> TokenStream
    where
        Self: Sized,
    { ... }
}

Types that can be interpolated inside a quote! invocation.

Required methods

fn to_tokens(&self, tokens: &mut TokenStream)

Write self to the given TokenStream.

Loading content...
Provided methods

fn to_token_stream(&self) -> TokenStream

Convert self directly into a TokenStream object.

This method is implicitly implemented using to_tokens, and acts as a convenience method for consumers of the ToTokens trait.

fn into_token_stream(self) -> TokenStream where
Self: Sized,

Convert self directly into a TokenStream object.

This method is implicitly implemented using to_tokens, and acts as a convenience method for consumers of the ToTokens trait.

Loading content...